Update patch set 1
Patch Set 1: Code-Review+1

As David noted, we can't drop those macros because of how the automated builds work.  What we need to do is determine the right upstream branching and release/tag structure to make things work for this package.  

Posting the comments from a separate mail thread:
It is a breaking change that might require some maneuvering to get it right. It wouldn’t work as is:

https://review.rdoproject.org/r/#/c/3643/

From Aditya:

Right now ‘upstream_version’ is taken from the ‘version' and ‘milestone’ of the openstack project. However, at BSN we had to modify it a bit to fit our needs. Instead of <version>.<milestone>.<bug-fix-id>, we have <version>.<bcf-controller-version>.<bug-fix-id>.

Since we manually update the version string every time we release a bug fix, the diff above would work. However, in your case, it wouldn’t have the 'bcf-version’ string to figure out the second part. I’m not sure how we can pass a new variable over there.

From me:

Ok, the way most openstack projects work is somewhat different. The way things normally work is through git tags and released tarballs.  

For example, if you want 1.2.3, you would create a 1.2.3 tag and push that to the github repo, then do a 1.2.3 release of the tarball on tarballs.openstack.org (following the same process every other openstack project follows.  

If you prefer to have different branches that that follow different number schemes (for example if you want all 1.2.x things on 1 branch, and 1.1.x things on a different branch, that's doable too.  In the RDO community, for the automated builds, we just need to override the branch being used.  So, if you want to support 1.2.x on Newton/OSP 10, we can pull from the 1.2.x branch (or whatever it's called.  If in ocata, you want 1.3.x instead, you would update to point there.  This is done through the rdo.yaml file in the rdoinfo repo[1].

Mike

[1] https://review.rdoproject.org/r/#/admin/projects/rdoinfo

Patch-set: 1