Update patch set 1
Patch Set 1:

Note that cbs build looks for tarball for 10.0.1, while the last version uploaded is 10.0.4. Could you update the version in the same change. Additionally, for proper versioning of dlrn build we'd need tag 10.0.4 pushed to stable/ocata. As this project is not using pbr for versioning it's important to get tags pushed when updating version in setup.cfg.

Patch-set: 1