Update patch set 1
Patch Set 1: Code-Review+2

It seems that when we pushed epoch 1, branches for stable releases did not exist so we added the epoch to old releases incorrectly.

Now we have semver for all supported releases:
mitaka -> 8.40.7
newton -> 9.42.0
ocata -> 10.0.5
master-> 10.0.0 (missing newer tag in master)

we'll need to backport that up to mitaka branch.

Patch-set: 1
Label: Code-Review=+2