cleanup redundant packages from queens and onwards
we had to split the networking-bigswitch package into multiple
packages owing to the containerization of services in queens.

we now install the packages based on requirement on a per
container basis - instead of the previous approach of installing
everything as part of networking-bigswitch.

now that all the containers are working with new packages, we
have to cleanup the networking-bigswitch and get rid of the
packages that are redundant.

handled in networking-bigswitch with the following change:
https://review.openstack.org/#/c/619096/

Change-Id: Ib001eda1ce184896114b9f784103d949b4e38758